zcash_client_sqlite: Move pinned table structures to `wallet::db` module
